// app/lead/search/component/LeadCard/LeadCard.module.scss.d.ts
export type Styles = {
  'borderBottomRound': string;
  'borderTopRound': string;
  'cardBase': string;
  'container': string;
  'interestedNewProject': string;
  'interestedNewProject_normal': string;
  'interestedNewProject_strong': string;
  'interestedNewProject_weak': string;
  'leadScore': string;
  'leadScore_cold': string;
  'leadScore_hot': string;
  'leadScore_warm': string;
  'line': string;
  'link': string;
  'minWidth': string;
  'showLessIcon': string;
  'showMoreIcon': string;
};

export type ClassNames = keyof Styles;

declare const styles: Styles;

export default styles;